Move template logic to a separate file
